+++ client/src/Pages/Home.js
@@ -5,29 +5,11 @@ import "./Home.css";
 import axios from "axios";
 import ImageGrid from "../ImageGrid";
 import receiptsApi from "../api/receiptsApi";
-import algoliasearch from "algoliasearch/lite";
-import { InstantSearch, SearchBox, Hits } from "react-instantsearch-dom";
-import DefaultImg from "../images/default-img.jpg";
 
-// search area
+import DefaultImg from "../images/default-img.jpg";
+import Vegas1 from "../images/vegas1.jpg";
 
 const API_URL = "http://localhost:8000";
-
-const searchClient = algoliasearch(
-  "FRJVR4E6Y4",
-  "c85d022bc27816c31277cfc57bc250ff"
-);
-
-const Hit = ({ hit }) => {
-  console.log("hit", hit);
-  return (
-    <div>
-      <p>
-        {hit.name} || {hit.category}
-      </p>
-    </div>
-  );
-};
 
 const Home = () => {
   const [userReceipts, setUserReceipts] = useState([]);
@@ -135,43 +117,44 @@ const Home = () => {
   return (
     <div>
       <div className="image-box">
-        <ImageGrid />
-        <div className="search-container">
-          <InstantSearch searchClient={searchClient} indexName="receipts">
-            <SearchBox />
-            <Hits hitComponent={Hit} />
-          </InstantSearch>
-        </div>
+        <img
+          src={Vegas1}
+          style={{ height: "320px", width: "1500px" }}
+          alt="coverphoto"
+        ></img>
       </div>
       <div className="receipt-results">
-        <table id="receipts">
-          <tr>
-            <th>Name</th>
-            <th>Category</th>
-            <th>Subtotal</th>
-            <th>Total</th>
-            <th>Date</th>
-          </tr>
-          {userReceipts.map((item, index) => {
-            return (
-              <tr className="receipt-individual">
-                <td> {item.name}</td>
-                <td> {item.category}</td>
-                <td>$ {item.subtotal}</td>
-                <td>$ {item.total}</td>
-                <td>$ {item.date}</td>
-                <button onClick={() => updateReceipt(item.id)}>Update</button>
-                <button
-                  index={item._id}
-                  onClick={() => deleteReceipt(item._id)}
-                  className="delete-button"
-                >
-                  X
-                </button>
-              </tr>
-            );
-          })}
-        </table>
+        <div className="title-section">
+          <h3>Manage Expenses</h3>
+          <table id="receipts">
+            <tr>
+              <th>Name</th>
+              <th>Category</th>
+              <th>Subtotal</th>
+              <th>Total</th>
+              <th>Date</th>
+            </tr>
+            {userReceipts.map((item, index) => {
+              return (
+                <tr className="receipt-individual">
+                  <td> {item.name}</td>
+                  <td> {item.category}</td>
+                  <td>$ {item.subtotal}</td>
+                  <td>$ {item.total}</td>
+                  <td>$ {item.date}</td>
+                  <button onClick={() => updateReceipt(item.id)}>Update</button>
+                  <button
+                    index={item._id}
+                    onClick={() => deleteReceipt(item._id)}
+                    className="delete-button"
+                  >
+                    X
+                  </button>
+                </tr>
+              );
+            })}
+          </table>
+        </div>
       </div>
       <div className="form-Container">
         <div className="receipt-form-container">
